requirements.txt
Django==4.0.6
django-filter==21.1
django-mysql==4.2.0
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
drf-spectacular==0.21.0
phonenumbers==8.12.37
drf-extra-fields==3.4.0
Pillow==8.4.0
mysqlclient
django-shortuuidfield
django-rest-knox==4.2.0
pytz